requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
werkzeug==2.3.7
faiss-cpu==1.7.4
langchain==0.2.11
langchain-community==0.2.10
langchain-groq==0.1.0
sentence-transformers==2.6.1
transformers==4.42.3
torch==2.3.1
huggingface-hub==0.23.1
numpy==1.26.4
requests==2.32.3
pdfplumber==0.11.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1